dom/tests/mochitest/crypto/test_legacy.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test presence of legacy window.crypto features when
         MOZ_DISABLE_CRYPTOLEGACY is NOT set.</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">

ok("crypto" in window, "crypto in window");
ok("version" in window.crypto, "version in window.crypto");
ok("enableSmartCardEvents" in window.crypto,
   "enableSmartCardEvents in window.crypto");
ok("generateCRMFRequest" in window.crypto,
   "generateCRMFRequest in window.crypto");
ok("importUserCertificates" in window.crypto,
   "importUserCertificates in window.crypto");
ok("popChallengeResponse" in window.crypto,
   "popChallengeResponse in window.crypto");
ok("random" in window.crypto, "random in window.crypto");
ok("signText" in window.crypto, "signText in window.crypto");
ok("disableRightClick" in window.crypto,
   "disableRightClick in window.crypto");

</script>
</body></html>
